refactor(control): split udp endpoint admission
